get, post, put and delete operations ready for user and event. Created event model. Restricted event creation to authenticated users. Restricted event deletion and update to event creators. Restricted user deletion and update to same user only

+++ controllers/users.js
@@ -2,10 +2,15 @@ const bcrypt = require('bcrypt')
 const usersRouter = require('express').Router()
 const User = require('../models/user')
 const middleware = require('../utils/middleware')
+const config = require('../utils/config')
 
 usersRouter.get('/', middleware.userExtractor, async (request, response) => {
-
-  const users = await User.find({})
+  const users = await User.find({}).populate('createdEvents', {
+    title: 1
+  })
+    .populate('respondedEvents', {
+      title: 1
+    })
   response.json(users)
 })
 
@@ -33,26 +38,66 @@ usersRouter.post('/', async (request, response) => {
   const user = new User({
     email: email,
     name: name,
-    passwordHash: passwordHash
+    passwordHash: passwordHash,
+    createdEvents: [],
+    respondedEvents: []
   })
 
   const savedUser = await user.save()
 
-  response.status(201).json(savedUser)
+  const userToReturn = await User.findById(savedUser._id).populate('createdEvents', {
+    title: 1
+  })
+    .populate('respondedEvents', {
+      title: 1
+    })
+
+  response.status(201).json(userToReturn)
 })
 
-usersRouter.delete( '/:id', async (request, response) => {
+usersRouter.delete( '/:id', middleware.userExtractor, async (request, response) => {
+
+  const user = await User.findById(request.params.id)
+
+  if (!user) {
+    return response.status(404).json({ error: 'user does not exist' })
+  }
+
+  const loggedUser = request.user
+
+  if (loggedUser._id.toString() !== user._id.toString()) {
+    return response
+      .status(401)
+      .json({ error: 'only the user can delete itself' })
+  }
+
+  await config.redisClient.del(request.params.id)
 
   await User.findByIdAndRemove(request.params.id)
   response.status(204).end()
 
 })
 
-usersRouter.put('/:id', async (request, response) => {
+usersRouter.put('/:id', middleware.userExtractor ,async (request, response) => {
   const body = request.body
 
+  const user = await User.findById(request.params.id)
+
+  if (!user) {
+    return response.status(404).json({ error: 'user does not exist' })
+  }
+
+  const loggedUser = request.user
+
+  if (loggedUser._id.toString() !== user._id.toString()) {
+    return response
+      .status(401)
+      .json({ error: 'only the user can modify itself' })
+  }
+
   const receivedUser = {
-    name: body.name
+    name: body.name,
+    respondedEvents: body.respondedEvents
   }
 
   const updatedUser = await User.findByIdAndUpdate(request.params.id, receivedUser, {
